moving image scripts to their own directory
