+++ lib/themes/dosomething/paraneue_dosomething/js/images/ImageUploadBeta.js
@@ -3,8 +3,8 @@ define(function(require) {
 
   var $         = require("jquery");
   var Modal     = require("neue/modal");
-  var ImageCrop = require("campaign/ImageCrop");
-  var Events = require("neue/events");
+  var ImageCrop = require("images/ImageCrop");
+  var Events    = require("neue/events");
 
   /**
    * Handles the modal interface for image upload buttons.
