+++ lib/themes/dosomething/paraneue_dosomething/js/reportback/Reportback.js
@@ -4,7 +4,7 @@ define(function(require) {
   var $                = require("jquery");
   var _                = require("lodash");
   var Events           = require("neue/events");
-  var ImageUploadBeta  = require("../campaign/ImageUploadBeta");
+  var ImageUploadBeta  = require("../images/ImageUploadBeta");
   var reportbackTplSrc = require("text!reportback/templates/reportback.tpl.html");
